# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CharacterResponse:
    properties:
      character:
        description: Display name of the character
        example: Ushiromiya Battler
        type: string
      characterId:
        description: Numeric character identifier
        example: "10"
        type: string
      limit:
        description: Maximum results per page
        example: 50
        type: integer
      offset:
        description: Pagination offset
        example: 0
        type: integer
      quotes:
        description: Paginated list of quotes
        items:
          type: object
        type: array
      total:
        description: Total number of matching quotes
        example: 4304
        type: integer
    type: object
  dto.ContextResponse:
    properties:
      after:
        description: Dialogue lines after the target quote
        items:
          type: object
        type: array
      before:
        description: Dialogue lines before the target quote
        items:
          type: object
        type: array
      quote:
        description: The target quote
        type: object
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        description: Error message
        example: query parameter 'q' is required
        type: string
    type: object
...
  dto.SearchAPIResponse:
    properties:
      limit:
        description: Maximum results per page
        example: 30
        type: integer
      offset:
        description: Pagination offset
        example: 0
        type: integer
      query:
        description: The search query that was used
        example: without love
        type: string
      results:
        description: Matching quotes with relevance scores
        items:
          $ref: '#/definitions/dto.SearchResult'
        type: array
      total:
        description: Total number of matches
        example: 25
        type: integer
    type: object
  dto.SearchResult:
    properties:
      quote:
        description: The matched quote
        type: object
      score:
        description: Relevance score (100 = exact match)
        example: 100
        type: integer
    type: object
...
info:
  contact:
    email: FAA@auaurora.moe
    name: Featherine Augustus Aurora
    url: https://x.com/FeatherineFAA
  description: API for searching and browsing quotes from Umineko no Naku Koro ni
    and Higurashi no Naku Koro ni
  title: When They Cry Quote API
  version: "1.0"
paths:
...
  /umineko/browse:
    get:
      description: Browse all quotes with optional filters and pagination
      parameters:
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      - description: Filter by character ID
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: character
        type: string
      - description: 'Interaction filter: first character ID (requires interactionB)'
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: interactionA
        type: string
      - description: 'Interaction filter: second character ID (requires interactionA)'
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: interactionB
        type: string
      - default: 50
        description: Maximum results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by episode (1-8)
        in: query
        name: episode
        type: integer
      - description: Filter by truth type
        enum:
        - red
        - blue
        - gold
        - purple
        in: query
        name: truth
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CharacterResponse'
      summary: Browse quotes
      tags:
      - umineko
...
  /umineko/context/{audioId}:
    get:
      description: Returns surrounding dialogue lines for a specific quote
      parameters:
      - description: Audio ID of the quote
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      - default: 5
        description: Number of context lines before and after
        in: query
        name: lines
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContextResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote context
      tags:
      - umineko
  /umineko/nearest-voiced/{audioId}:
    get:
      description: Returns the audio ID of the nearest quote with voice audio in a
        given direction
      parameters:
      - description: Audio ID of the reference quote
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      - default: next
        description: Direction to search
        enum:
        - next
        - prev
        in: query
        name: direction
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Find nearest voiced quote
      tags:
      - umineko
  /umineko/quote/{audioId}:
    get:
      description: Returns a specific quote identified by its audio ID
      parameters:
      - description: Audio ID of the quote
        in: path
        name: audioId
        required: true
        type: string
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.UminekoQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote by audio ID
      tags:
      - umineko
  /umineko/quote/index/{index}:
    get:
      description: Returns a specific quote identified by its position index in the
        parsed script
      parameters:
      - description: Index of the quote in the parsed script
        in: path
        name: index
        required: true
        type: integer
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.UminekoQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get quote by index
      tags:
      - umineko
  /umineko/random:
    get:
      description: Returns a random quote with optional filters
      parameters:
      - default: en
        description: Language
        enum:
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      - description: Filter by character ID
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: character
        type: string
      - description: Filter by episode (1-8)
        in: query
        name: episode
        type: integer
      - description: Filter by truth type
        enum:
        - red
        - blue
        - gold
        - purple
        in: query
        name: truth
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/umineko_quote_internal_dto.UminekoQuote'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get random quote
      tags:
      - umineko
  /umineko/search:
    get:
      description: Search for quotes by text query with optional filters
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: en
        description: Language
        enum:
        - auto
        - en
        - wh
        - ja
        - ru
        - es
        - pt
        in: query
        name: lang
        type: string
      - default: 30
        description: Maximum results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - description: Filter by character ID
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: character
        type: string
      - description: 'Interaction filter: first character ID (requires interactionB)'
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: interactionA
        type: string
      - description: 'Interaction filter: second character ID (requires interactionA)'
        enum:
        - group_voices
        - kinzo
        - krauss
        - natsuhi
        - jessica
        - eva
        - hideyoshi
        - george
        - rudolf
        - kyrie
        - battler
        - ange
        - rosa
        - maria
        - genji
        - shannon
        - kanon
        - gohda
        - kumasawa
        - nanjo
        - amakusa
        - okonogi
        - kasumi
        - professor
        - kawabata
        - nanjo_son
        - kumasawa_son
        - beatrice
        - bernkastel
        - lambdadelta
        - virgilia
        - ronove
        - gaap
        - sakutarou
        - eva_beatrice
        - chiester_45
        - chiester_410
        - chiester_00
        - lucifer
        - leviathan
        - satan
        - belphegor
        - mammon
        - beelzebub
        - asmodeus
        - goat
        - erika
        - dlanor
        - gertrude
        - cornelia
        - featherine
        - zepar
        - furfur
        - lion
        - will
        - clair
        - ikuko
        - tohya
        - kinzo_young
        - bice
        - beato_elder
        - misc_voices
        - narrator
        in: query
        name: interactionB
        type: string
      - description: Filter by episode (1-8)
        in: query
        name: episode
        type: integer
      - description: Filter by truth type
        enum:
        - red
        - blue
        - gold
        - purple
        in: query
        name: truth
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SearchAPIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Search quotes
      tags:
      - umineko
  /umineko/stats:
    get:
      description: Returns statistics about quotes including top speakers, truth per
        episode, and character interactions
      parameters:
      - description: Filter by episode (1-8), 0 for all
        in: query
        name: episode
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatsResult'
      summary: Get quote statistics
      tags:
      - umineko
schemes:
- https
- http
swagger: "2.0"